Change font, layout, and navigation menu of home page
The new font gives a less 'fantasy' feel, but retains the'child and jolly'
mood.

Delete the newsletter section in the middle of the body because it looks
better, and to keep this website static in order to be able to host on
Git lab.

Add two new files for the amended navigation menu in the header.